+++ API/CarometroAPI/Domains/Professor.cs
@@ -7,8 +7,8 @@ namespace CarometroAPI.Domains
 {
     public partial class Professor
     {
-        public short IdProfessor { get; set; }
-        public short? IdUsuario { get; set; }
+        public int IdProfessor { get; set; }
+        public int? IdUsuario { get; set; }
         public string Matricula { get; set; }
 
         public virtual Usuario IdUsuarioNavigation { get; set; }
